chore: move auth-jwt
