+++ packages/nestjs-auth-apple/src/auth-apple.module.spec.ts
@@ -1,8 +1,10 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { TypeOrmExtModule } from '@concepta/nestjs-typeorm-ext';
-import { AuthenticationModule } from '@concepta/nestjs-authentication';
-import { AuthJwtModule } from '@concepta/nestjs-auth-jwt';
-import { JwtModule } from '@concepta/nestjs-authentication';
+import {
+  AuthenticationModule,
+  AuthJwtModule,
+  JwtModule,
+} from '@concepta/nestjs-authentication';
 import { CrudModule } from '@concepta/nestjs-crud';
 import {
   UserModule,
