+++ packages/nestjs-auth-google/package.json
@@ -24,7 +24,7 @@
     "passport-google-oauth20": "^2.0.0"
   },
   "devDependencies": {
-    "@concepta/nestjs-auth-jwt": "^6.0.0-alpha.3",
+    "@concepta/nestjs-authentication": "^6.0.0-alpha.3",
     "@concepta/nestjs-crud": "^6.0.0-alpha.3",
     "@concepta/nestjs-password": "^6.0.0-alpha.3",
     "@concepta/nestjs-typeorm-ext": "^6.0.0-alpha.3",
