--- packages/nestjs-auth-jwt/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "@concepta/nestjs-auth-jwt",
-  "version": "6.0.0-alpha.3",
-  "description": "Rockets NestJS JWT Authorization",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "license": "BSD-3-Clause",
-  "publishConfig": {
-    "access": "public"
-  },
-  "files": [
-    "dist/**/!(*.spec|*.e2e-spec|*.fixture).{js,d.ts}"
-  ],
-  "dependencies": {
-    "@concepta/nestjs-authentication": "^6.0.0-alpha.3",
-    "@concepta/nestjs-common": "^6.0.0-alpha.3",
-    "@concepta/nestjs-exception": "^6.0.0-alpha.3",
-    "@concepta/typeorm-common": "^6.0.0-alpha.3",
-    "@nestjs/common": "^10.4.1",
-    "@nestjs/config": "^3.2.3",
-    "@nestjs/core": "^10.4.1"
-  },
-  "devDependencies": {
-    "@nestjs/testing": "^10.4.1",
-    "jest-mock-extended": "^2.0.9"
-  },
-  "peerDependencies": {
-    "class-validator": "*",
-    "rxjs": "^7.1.0",
-    "typeorm": "^0.3.0"
-  }
-}
