--- packages/nestjs-auth-jwt/src/interfaces/auth-jwt-settings.interface.ts
@@ -1,3 +1,0 @@
-import { JwtStrategyOptionsInterface } from '@concepta/nestjs-authentication';
-
-export interface AuthJwtSettingsInterface extends JwtStrategyOptionsInterface {}
