--- packages/nestjs-auth-jwt/tsconfig.json
@@ -1,15 +1,0 @@
-{
-  "extends": "../../tsconfig",
-  "compilerOptions": {
-    "composite": true,
-    "rootDir": "./src",
-    "outDir": "./dist",
-    "typeRoots": [
-      "./node_modules/@types",
-      "../../node_modules/@types"
-    ]
-  },
-  "include": [
-    "src/**/*.ts"
-  ]
-}
