--- packages/nestjs-auth-jwt/typedoc.json
@@ -1,3 +1,0 @@
-{
-  "entryPoints": ["src/index.ts"]
-}
