+++ packages/nestjs-auth-local/package.json
@@ -24,7 +24,7 @@
     "passport-local": "^1.0.0"
   },
   "devDependencies": {
-    "@concepta/nestjs-auth-jwt": "^6.0.0-alpha.3",
+    "@concepta/nestjs-authentication": "^6.0.0-alpha.3",
     "@nestjs/testing": "^10.4.1",
     "@types/passport-local": "^1.0.38",
     "@types/supertest": "^2.0.16",
