+++ packages/nestjs-auth-recovery/package.json
@@ -21,7 +21,6 @@
     "@nestjs/swagger": "^7.4.0"
   },
   "devDependencies": {
-    "@concepta/nestjs-auth-jwt": "^6.0.0-alpha.3",
     "@concepta/nestjs-authentication": "^6.0.0-alpha.3",
     "@concepta/nestjs-crud": "^6.0.0-alpha.3",
     "@concepta/nestjs-email": "^6.0.0-alpha.3",
