+++ packages/nestjs-auth-verify/src/__fixtures__/app.module.fixture.ts
@@ -1,8 +1,10 @@
 import { Module } from '@nestjs/common';
 import { EmailModule, EmailService } from '@concepta/nestjs-email';
-import { AuthenticationModule } from '@concepta/nestjs-authentication';
-import { AuthJwtModule } from '@concepta/nestjs-auth-jwt';
-import { JwtModule } from '@concepta/nestjs-authentication';
+import {
+  AuthenticationModule,
+  AuthJwtModule,
+  JwtModule,
+} from '@concepta/nestjs-authentication';
 
 import { AuthVerifyModule } from '../auth-verify.module';
