+++ packages/nestjs-authentication/package.json
@@ -14,6 +14,7 @@
   "dependencies": {
     "@concepta/nestjs-common": "^6.0.0-alpha.3",
     "@concepta/nestjs-exception": "^6.0.0-alpha.3",
+    "@concepta/typeorm-common": "^6.0.0-alpha.3",
     "@nestjs/common": "^10.4.1",
     "@nestjs/config": "^3.2.3",
     "@nestjs/core": "^10.4.1",
@@ -36,6 +37,8 @@
   },
   "peerDependencies": {
     "class-transformer": "*",
-    "class-validator": "*"
+    "class-validator": "*",
+    "rxjs": "^7.1.0",
+    "typeorm": "^0.3.0"
   }
 }
