+++ packages/nestjs-authentication/src/__fixtures__/user/user-lookup.service.fixture.ts
@@ -3,7 +3,7 @@ import {
   ReferenceIdInterface,
   ReferenceSubject,
 } from '@concepta/nestjs-common';
-import { AuthJwtUserLookupServiceInterface } from '../../interfaces/auth-jwt-user-lookup-service.interface';
+import { AuthJwtUserLookupServiceInterface } from '../../auth-jwt/interfaces/auth-jwt-user-lookup-service.interface';
 
 @Injectable()
 export class UserLookupServiceFixture
