+++ packages/nestjs-authentication/src/__fixtures__/user/user.controller.fixture.ts
@@ -1,5 +1,5 @@
 import { Controller, Get, UseGuards } from '@nestjs/common';
-import { AuthJwtGuard } from '../../auth-jwt.guard';
+import { AuthJwtGuard } from '../../auth-jwt/auth-jwt.guard';
 
 @Controller('user')
 @UseGuards(AuthJwtGuard)
