+++ packages/nestjs-authentication/src/__fixtures__/user/user.entity.fixture.ts
