+++ packages/nestjs-authentication/src/auth-jwt/auth-jwt.module-definition.spec.ts
