+++ packages/nestjs-authentication/src/auth-jwt/config/auth-jwt-default.config.ts
@@ -1,5 +1,5 @@
 import { registerAs } from '@nestjs/config';
-import { ExtractJwt } from '@concepta/nestjs-authentication';
+import { ExtractJwt } from '../../jwt/index';
 import { AUTH_JWT_MODULE_DEFAULT_SETTINGS_TOKEN } from '../auth-jwt.constants';
 import { AuthJwtSettingsInterface } from '../interfaces/auth-jwt-settings.interface';
