+++ packages/nestjs-authentication/src/auth-jwt/exceptions/auth-jwt.exception.ts
