+++ packages/nestjs-authentication/src/auth-jwt/interfaces/auth-jwt-options-extras.interface.ts
