+++ packages/nestjs-authentication/src/auth-jwt/interfaces/auth-jwt-options.interface.ts
@@ -1,7 +1,8 @@
 import { CanActivate } from '@nestjs/common';
-import { VerifyTokenServiceInterface } from '@concepta/nestjs-authentication';
+
 import { AuthJwtSettingsInterface } from './auth-jwt-settings.interface';
 import { AuthJwtUserLookupServiceInterface } from './auth-jwt-user-lookup-service.interface';
+import { VerifyTokenServiceInterface } from '../../core/interfaces/verify-token-service.interface';
 
 export interface AuthJwtOptionsInterface {
   settings?: AuthJwtSettingsInterface;
