+++ packages/nestjs-authentication/src/auth-jwt/interfaces/auth-jwt-settings.interface.ts
@@ -1,0 +1,3 @@
+import { JwtStrategyOptionsInterface } from '../../jwt/interfaces/jwt-strategy-options.interface';
+
+export interface AuthJwtSettingsInterface extends JwtStrategyOptionsInterface {}
