+++ packages/nestjs-authentication/src/auth-jwt/interfaces/auth-jwt-user-lookup-service.interface.ts
