+++ packages/nestjs-samples/package.json
@@ -12,10 +12,9 @@
     "dist/**/!(*.spec|*.e2e-spec|*.fixture).{js,d.ts}"
   ],
   "dependencies": {
-    "@concepta/nestjs-auth-jwt": "^6.0.0-alpha.3",
+    "@concepta/nestjs-authentication": "^6.0.0-alpha.3",
     "@concepta/nestjs-auth-local": "^6.0.0-alpha.3",
     "@concepta/nestjs-auth-refresh": "^6.0.0-alpha.3",
-    "@concepta/nestjs-authentication": "^6.0.0-alpha.3",
     "@concepta/nestjs-common": "^6.0.0-alpha.3",
     "@concepta/nestjs-crud": "^6.0.0-alpha.3",
     "@concepta/nestjs-email": "^6.0.0-alpha.3",
