+++ packages/nestjs-samples/src/03-authentication/app.module.ts
@@ -2,10 +2,12 @@ import { Module } from '@nestjs/common';
 import { TypeOrmExtModule } from '@concepta/nestjs-typeorm-ext';
 import { AuthLocalModule } from '@concepta/nestjs-auth-local';
 import { AuthRefreshModule } from '@concepta/nestjs-auth-refresh';
-import { AuthJwtModule } from '@concepta/nestjs-auth-jwt';
-import { AuthenticationModule } from '@concepta/nestjs-authentication';
+import {
+  AuthenticationModule,
+  AuthJwtModule,
+  JwtModule,
+} from '@concepta/nestjs-authentication';
 import { UserModule, UserLookupService } from '@concepta/nestjs-user';
-import { JwtModule } from '@concepta/nestjs-authentication';
 import { PasswordModule } from '@concepta/nestjs-password';
 import { CrudModule } from '@concepta/nestjs-crud';
 import { CustomUserController } from './user/user.controller';
