+++ packages/nestjs-user/src/__fixtures__/app.module.custom.fixture.ts
@@ -2,9 +2,11 @@ import { AccessControl } from 'accesscontrol';
 import { Module } from '@nestjs/common';
 import { TypeOrmExtModule } from '@concepta/nestjs-typeorm-ext';
 import { PasswordModule } from '@concepta/nestjs-password';
-import { AuthenticationModule } from '@concepta/nestjs-authentication';
-import { AuthJwtModule } from '@concepta/nestjs-auth-jwt';
-import { JwtModule } from '@concepta/nestjs-authentication';
+import {
+  AuthenticationModule,
+  AuthJwtModule,
+  JwtModule,
+} from '@concepta/nestjs-authentication';
 import { AccessControlModule } from '@concepta/nestjs-access-control';
 import { CrudModule } from '@concepta/nestjs-crud';
 import { EventModule } from '@concepta/nestjs-event';
