+++ packages/nestjs-user/src/__fixtures__/app.module.fixture.ts
@@ -1,10 +1,12 @@
 import { AccessControl } from 'accesscontrol';
 import { Module } from '@nestjs/common';
 import { TypeOrmExtModule } from '@concepta/nestjs-typeorm-ext';
-import { AuthenticationModule } from '@concepta/nestjs-authentication';
+import {
+  AuthenticationModule,
+  AuthJwtModule,
+  JwtModule,
+} from '@concepta/nestjs-authentication';
 import { AccessControlModule } from '@concepta/nestjs-access-control';
-import { JwtModule } from '@concepta/nestjs-authentication';
-import { AuthJwtModule } from '@concepta/nestjs-auth-jwt';
 import { PasswordModule } from '@concepta/nestjs-password';
 import { CrudModule } from '@concepta/nestjs-crud';
 import { EventModule } from '@concepta/nestjs-event';
